+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - New #26: Add `copyMessage` and `copyMessages` methods.
 - New #27: Add `sendAnimation` method.
 - New #28: Add `sendVoice` method.
+- New #29: Add `sendVideoNote` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.